Move css out of css folder to their corrosponding component
