+++ client/src/components/Admin/IssueForm/Alternative.css
@@ -54,5 +54,5 @@
 }
 
 .title {
-  composes: title from './IssueForm.css';
+  composes: title from './index.css';
 }
