+++ client/src/components/Alternative.js
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
-import css from '../css/Alternative.css';
+import css from './Alternative.css';
 
 const Alternative = ({ disabled, id, text, selected, onClick }) => {
   const alternativeClass = classNames(css.alternative, {
