+++ client/src/components/Alternatives.css
