+++ client/src/components/App/Issue.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { getIssueText } from '../../selectors/issues';
 import Card from '../Card';
-import css from '../../css/Issue.css';
+import css from './Issue.css';
 
 const Issue = ({ text }) => (
   <Card
