+++ client/src/components/ConcludedIssue.css
