+++ client/src/components/ConcludedIssueList.css
