+++ client/src/components/Heading.js
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import { Link } from 'react-router';
-import css from '../css/Heading.css';
+import css from './Heading.css';
 
 const Heading = ({ link, title, children }) => (
   <header className={css.heading}>
