+++ client/src/components/IssueStatus.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { VoteStatusContainer } from './VoteStatus';
 import Card from './Card';
-import css from '../css/IssueStatus.css';
+import css from './IssueStatus.css';
 
 const IssueStatus = () => (
   <Card classes={css.status} subtitle="Foreløpig status">
