+++ client/src/components/VoteCounter.css
