+++ client/src/components/VoteStatus.css
